feat: probabilidades reales + fp16 + permalink y pulido UX
- pipeline.js: softmax(z/T) sobre el vocabulario completo — los % mostrados
  son las probabilidades verdaderas del modelo, no renormalizadas al top-k;
  el corte top-p conserva la semántica secuencial de HF (sin cambio de
  comportamiento en el muestreo)
- models.js: DistilGPT-2 y GPT-2 en fp16 (156MB vs 313MB, sin ruido de
  cuantización; verificado sobre WASM)
- main.js: permalink ?p= (prompt compartible, auto-arranque sin scroll)
  + guard isComposing en Enter
- muestreo.js: greedy deshabilita «Muestrear» y lo explica en el caption
- content.js: nota de embeddings posicionales (wpe) en Profundizar §2 +
  párrafo sobre probabilidades reales en §4
- style.css: rail móvil como píldora horizontal inferior (<900px) +
  .btn:disabled generalizado
- README/CLAUDE.md: tamaños actualizados y patrón de probabilidades

+++ js/models.js
@@ -25,7 +25,9 @@ export const MODEL_CONFIGS = {
     vocab_size: 50257,
     params: '82M',
     heads: 12,
-    dtype: 'fp32',
+    // fp16 = mitad de descarga que fp32 (156MB vs 313MB) sin ruido de
+    // cuantización; q8 solo baja a 226MB porque los embeddings quedan fp32.
+    dtype: 'fp16',
   },
   'onnx-community/gpt2-ONNX': {
     name: 'GPT-2',
@@ -35,7 +37,7 @@ export const MODEL_CONFIGS = {
     vocab_size: 50257,
     params: '124M',
     heads: 12,
-    dtype: 'fp32',
+    dtype: 'fp16',
   },
   'onnx-community/gpt2-medium-ONNX': {
     name: 'GPT-2 Medium',
